[Add] cleanup callback for Rigidbody component
